Redid everything to use annotations for ty_map.
Also replaced OptCons w/ Nil and SomeCons, and got rid of
Type::Nil, replacing it with the more general Type::Unresolved, for
which a lone `nil' simply has type `Type::Opt(Type::Unresolved)'

This commit also deleted a ton of code that will be rewritten using
(and designing) the new annotations and passes model. So the
compiler is a lot less functional right now than it was 2 commits
ago. This is a messy commit, but the next few commits which pull
things together should make it clear what's going on.